+++ src/main/java/lotto/domain/Store.java
@@ -1,22 +1,21 @@
 package lotto.domain;
 
-import lotto.domain.lottogenerator.LottoGenerator;
-
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class Store {
 
     private final List<LottoTicket> lottoTickets;
 
-    public Store(final LottoGenerator lottoGenerator) {
-        this.lottoTickets = lottoGenerator.create();
+    public Store(final List<LottoTicket> lottoTickets) {
+        this.lottoTickets = lottoTickets;
     }
 
-    public Store(final List<Store> lottoTickets) {
-        this.lottoTickets = lottoTickets.stream()
-                .flatMap(lottoTicket -> lottoTicket.getLottoTickets().stream()).collect(Collectors.toList());
+    public Store add(final Store store) {
+        return new Store(Stream.concat(this.lottoTickets.stream(), store.lottoTickets.stream())
+                .collect(Collectors.toList()));
     }
 
     public List<Rank> match(final WinningLotto winningLotto) {
