--- src/main/java/lotto/domain/lottogenerator/LottoGenerator.java
@@ -1,12 +1,0 @@
-package lotto.domain.lottogenerator;
-
-import lotto.domain.LottoTicket;
-
-import java.util.List;
-
-@FunctionalInterface
-public interface LottoGenerator {
-
-    List<LottoTicket> create();
-
-}
